current progress reindexing service architecture for indexing
